chore: Move `sequence_writer<T>` to the `sequence.h` header file
