+++ src/mjolnir/graphbuilder.cc
@@ -20,7 +20,6 @@
 #include "midgard/pointll.h"
 #include "midgard/polyline2.h"
 #include "midgard/sequence.h"
-#include "midgard/sequence_writer.h"
 #include "midgard/tiles.h"
 #include "midgard/util.h"
 #include "mjolnir/admin.h"
